# docs/swagger.yaml
definitions:
  APIError:
    properties:
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  APIResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  ChatActionReq:
    properties:
      jid:
        type: string
    required:
    - jid
    type: object
  ChatMarkReadReq:
    properties:
      jid:
        type: string
      messageIds:
        items:
          type: string
        type: array
    required:
    - jid
    - messageIds
    type: object
  ChatMarkUnreadReq:
    properties:
      jid:
        type: string
    required:
    - jid
    type: object
...
info:
  contact:
    name: wzap Support
    url: https://github.com/felipyfgs/wzap
  description: WhatsApp Multi-Session API powered by whatsmeow
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: wzap API
  version: "1.0"
paths:
...
  /sessions/{sessionId}/chat/archive:
    post:
      consumes:
      - application/json
      description: Archives a chat identified by JID, moving it out of the main chat
        list
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Chat JID payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ChatActionReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Archive a chat
      tags:
      - Chat
  /sessions/{sessionId}/chat/delete:
    post:
      consumes:
      - application/json
      description: Permanently deletes a chat (including media)
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Chat JID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ChatActionReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Delete a chat
      tags:
      - Chat
  /sessions/{sessionId}/chat/mute:
    post:
      consumes:
      - application/json
      description: Mutes notifications for a chat identified by JID
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Chat JID payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ChatActionReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Mute a chat
      tags:
      - Chat
  /sessions/{sessionId}/chat/pin:
    post:
      consumes:
      - application/json
      description: Pins a chat to the top of the chat list
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Chat JID payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ChatActionReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Pin a chat
      tags:
      - Chat
  /sessions/{sessionId}/chat/read:
    post:
      consumes:
      - application/json
      description: Marks specific messages in a chat as read
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Chat JID + message IDs
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ChatMarkReadReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Mark chat as read
      tags:
      - Chat
  /sessions/{sessionId}/chat/unarchive:
    post:
      consumes:
      - application/json
      description: Removes a chat from the archived list
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Chat JID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ChatActionReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Unarchive a chat
      tags:
      - Chat
  /sessions/{sessionId}/chat/unmute:
    post:
      consumes:
      - application/json
      description: Removes mute from a chat
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Chat JID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ChatActionReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Unmute a chat
      tags:
      - Chat
  /sessions/{sessionId}/chat/unpin:
    post:
      consumes:
      - application/json
      description: Removes a chat from the pinned position at the top of the chat
        list
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Chat JID payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ChatActionReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Unpin a chat
      tags:
      - Chat
  /sessions/{sessionId}/chat/unread:
    post:
      consumes:
      - application/json
      description: 'Marks a chat as unread. Note: WhatsApp protocol does not support
        this operation directly.'
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Chat JID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ChatMarkUnreadReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "501":
          description: Not implemented by WhatsApp protocol
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Mark chat as unread
      tags:
      - Chat
...
securityDefinitions:
  Authorization:
    description: Session or admin token (without Bearer prefix)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"